[REM] sort,selection: remove duplicated code
The code to expand a zone until there is only empty cells beside it
was duplicated in both `sort.ts` and `selection_stream_processor.ts`.

This commit removes the duplicated code and uses the one in
`selection_stream_processor.ts`. since it's simpler and shorter.
